fix(api/lyrics): compute blocked_by_asr_gap in all-stale endpoint
post_reprocess_all_stale hard-coded `blocked_by_asr_gap: 0`, so the
dashboard's #98 banner never fired on the "Reprocess all stale" path
even when N rows were quarantined under the asr_gap escape hatch.
Operators saw `queued: M` with no signal that other stale work was
unreachable until a future pipeline-version bump.

Compute the count via
  SELECT COUNT(*) FROM videos
  WHERE has_lyrics = 0 AND lyrics_source = 'asr_gap'
        AND lyrics_pipeline_version < ?
matching the scope of the UPDATE that follows.

Also revert e2e/mock-api.mjs's intentional hard-coded 0 — both
endpoints now share the same response shape, so the mock exposes a
second slot the __mock admin can drive via `target: "all-stale"`.

Sibling-split api/lyrics.rs into lyrics.rs (527 lines, handlers only)
and lyrics_tests.rs (474 lines, tests). The unified file hit the
1000-line cap after this change; the split keeps both well under and
matches the worker/orchestrator/mod_tests pattern used elsewhere.

Closes #101
